autoagentstudioapp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
pydantic==2.5.0
google-generativeai==0.4.0
python-dotenv==1.0.0
itsdangerous==2.2.0
sqlalchemy==2.0.48
mysql-connector-python==9.6.0
python-multipart==0.0.22
